# perf: optimize protocol validation in `ValidateProtocol`